Rut was changed from int to numeric

+++ CASO 2/caso2.sql
@@ -3,6 +3,6 @@ CREATE DATABASE modelo_dos;
 CREATE TABLE alumno(
     id SERIAL PRIMARY KEY,
     nombre VARCHAR(70) NOT NULL,
-    rut INT NOT NULL,
+    rut NUMERIC NOT NULL,
     curso INT NOT NULL FOREIGN KEY
 )
